requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# HTTP Client (with browser fingerprint simulation)
curl-cffi>=0.6.0

# Template Engine
jinja2>=3.1.3

# JWT Token Processing
pyjwt>=2.8.0

# Environment Variables
python-dotenv>=1.0.0

# Password Hashing
bcrypt>=4.1.2

# Encryption
cryptography>=41.0.0

# Data Validation
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-multipart>=0.0.6
itsdangerous>=2.1.2
pytz>=2023.3

# Excel Export
xlsxwriter>=3.1.9
